reorganize creation methods
